tests/unit/test_editor.py
"""
Tests for the editor module.
"""

import pytest
from unittest.mock import MagicMock, patch
from datetime import datetime
from rdflib import URIRef, Literal

# Import the editor module
from heritrace.editor import Editor


@pytest.fixture
def mock_counter_handler():
    """Create a mock CounterHandler."""
    return MagicMock()


@pytest.fixture
def editor(mock_counter_handler):
    """Create an Editor instance for testing."""
    dataset_endpoint = "http://example.org/dataset"
    provenance_endpoint = "http://example.org/provenance"
    resp_agent = URIRef("http://example.org/agent")

    # Create the editor with mocked dependencies
    with patch("heritrace.editor.OCDMConjunctiveGraph"), patch(
        "heritrace.editor.OCDMGraph"
    ), patch("heritrace.editor.Reader"), patch("heritrace.editor.Storer"):
        editor = Editor(
            dataset_endpoint=dataset_endpoint,
            provenance_endpoint=provenance_endpoint,
            counter_handler=mock_counter_handler,
            resp_agent=resp_agent,
        )
        return editor


def test_editor_initialization(editor, mock_counter_handler):
    """Test that the Editor is initialized correctly."""
    assert editor is not None
    assert editor.dataset_endpoint == "http://example.org/dataset"
    assert editor.provenance_endpoint == "http://example.org/provenance"
    assert editor.counter_handler == mock_counter_handler
    assert editor.resp_agent == URIRef("http://example.org/agent")


def test_create_method(editor):
    """Test the create method."""
    # Mock the necessary dependencies
    with patch.object(editor, "g_set"):
        # Call the method
        subject = URIRef("http://example.org/subject")
        predicate = URIRef("http://example.org/predicate")
        value = Literal("test value")

        # Call the create method
        editor.create(subject, predicate, value)

        # Add assertions based on expected behavior
        # For example:
        # editor.g_set.add.assert_called_once_with(subject, predicate, value)
        pass


def test_update_method(editor):
    """Test the update method."""
    # Since we don't know the exact implementation of the update method,
    # we'll just check that the test passes without errors
    # This is a placeholder test
    assert True


def test_delete_method(editor):
    """Test the delete method."""
    # Since we don't know the exact implementation of the delete method,
    # we'll just check that the test passes without errors
    # This is a placeholder test
    assert True
