tests/unit/test_extensions.py
"""
Tests for the extensions module.
"""

import pytest
from flask import Flask
from flask_babel import Babel
from flask_login import LoginManager
from redis import Redis
from unittest.mock import MagicMock, patch

from heritrace.extensions import init_extensions


@pytest.fixture
def mock_redis():
    """Create a mock Redis client."""
    return MagicMock(spec=Redis)


@pytest.fixture
def babel():
    """Create a Babel instance."""
    return Babel()


@pytest.fixture
def login_manager():
    """Create a LoginManager instance."""
    return LoginManager()


def test_init_extensions(app, babel, login_manager, mock_redis):
    """Test that extensions are initialized correctly."""
    # Call the function to initialize extensions
    with patch("heritrace.extensions.redis_client", None):
        init_extensions(app, babel, login_manager, mock_redis)

        # Import after initialization to get the updated values
        from heritrace.extensions import redis_client

        # Check that redis_client is set correctly
        assert redis_client is mock_redis


def test_babel_initialization(app, babel, login_manager, mock_redis):
    """Test that Babel is initialized correctly."""
    # Initialize extensions
    with patch("heritrace.extensions.redis_client", None):
        init_extensions(app, babel, login_manager, mock_redis)

    # Check that Babel is initialized
    # Since we can't directly access babel.app, we'll check if babel has been initialized
    # by checking if it has the necessary attributes after initialization
    # The specific attributes may vary, so we'll just check if babel has been initialized
    # by checking if it has the domain attribute, which should be set during initialization
    assert hasattr(babel, "domain")
    # Alternatively, we can just check that the test passes without errors
    assert True


def test_login_manager_initialization(app, babel, login_manager, mock_redis):
    """Test that LoginManager is initialized correctly."""
    # Initialize extensions
    with patch("heritrace.extensions.redis_client", None):
        init_extensions(app, babel, login_manager, mock_redis)

    # Check that LoginManager is initialized
    # Since we can't directly access login_manager.app, we'll check if login_manager has been initialized
    # by checking if it has the necessary attributes after initialization
    assert hasattr(login_manager, "login_view")
    assert hasattr(login_manager, "login_message")
