pyproject.toml
[tool.poetry]
name = "heritrace"
version = "0.1.0"
description = "An environment for data modification"
authors = ["arcangelo7 <arcangelomas@gmail.com>"]
license = "ISC"
readme = "README.md"
packages = [{include = "heritrace"}]

[tool.poetry.dependencies]
python = "^3.10,<3.14"
flask = "^2.3.3"
sparqlwrapper = "^2.0.0"
pyyaml = "^6.0.1"
flask-babel = "^3.1.0"
validators = "0.20.0"
requests = "^2.31.0"
time-agnostic-library = "^4.6.13"
flask-wtf = "^1.1.1"
oc-ocdm = "^9.0.0"
pyshacl = "0.23.0"
setuptools = "^68.2.2"
docker = "^7.1.0"
requests-oauthlib = "^2.0.0"
flask-login = "^0.6.3"
rdflib-ocdm = "^0.4.1"
pyopenssl = "^24.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
coverage = "^7.6.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--junitxml=pytest-report.xml --cov=heritrace --cov-report=xml"
